Properly search for PipeWire libs
Summary:
Recent change in upcoming PipeWire releases made all headers to be
installed versioned directories so multiple versions can be installed
at once. We have to search a bit differently now. This change makes
it to search first for PipeWire 0.3 and if it's not found, we try to
search for the previous version.

Reviewers: #plasma, cgiboudeaux

Reviewed By: cgiboudeaux

Subscribers: cgiboudeaux, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D27277