Pre-select the only item in the list lists with single items
When an app asks for permission to see the screen but there is only a
single screen, or permission to see a window but there is only a single
window, the user will be shown a permissions dialog that has a list view
of screens or windows with only one item in it. The user will need to
manually click on that one item before the Share button will become
enabled, which is annoying and somewhat unintuitive.

This commit refines that workflow so that if the permissions dialog is
showing a single item in its list, that single item is pre-selected so
that the user only has to click on the "Share" button.


(cherry picked from commit e67730fb658c425a74f3e559f625c44c50ab1415)