make sure host okular shows up
when using (e.g.) dolphin as a flatpak it runs stuff through the app
chooser dialog but that dialog is incapable of handling okular
correctly. okular is a bit of an odd ball because it installs a separate
desktop file for every file format it supports (thereby technically
allowing default handling). unfortunately since currently portals always
prompt which app to use (until a clear winner choice has been made) this
system falls flat on the face because now we somehow need to get okular
into our preference list but the actual preference we get passed in from
the upstream portal is 'okularApplication_pdf' not 'org.kde.okular'.
Very unfortunate set of circumstances that. To mitigate this problem we
use the new aliasfor property to compile a list of "fake" names and then
present the aliased entity instead of the fake one.

this is not a problem when okular is installed as a flatpak because
there the mimetype list of the flatpak's desktop file is comprehensive