AppChooserDialog: set title and mainText separately
Right now this code is making use of a bug in SystemDialog.qml which
causes the title property to be duplicated across both the window title
and also the main text in the dialog itself. This is now fixed in
SystemDialog, so now let's use the API properly and set the title and
mainText properties separately to improve the presentation of the
dialog.