RemoteDesktop: Simplify interaction to a minimum
It doesn't make any sense for the user to be nitpicky about the
configuration of a RemoteDesktop session. Ask users if they agree to the
sharing that was requested by the app or not, there is little reason to
decide there.

For the outputs, share all outputs if the app can take several streams,
offer a big composite stream otherwise.