screensharing, RemoteDesktop: Move SNI to the session
This way non-screensharing sessions will also be tracked by a UI
element. Also this means that instead of closing the pipewire stream we
close the session which also feels more correct, since we are
encapsulating the SNI in a session object rather than making the object
work on the spot and guess its state through random signals.