Background: Make allow and forbid actions more discoverable
"Allow" and "Force quit" actions are buried in a dialog which is
opened by clicking the "Find out more" button, which is deceptive.

The find out more dialog asks the user whether they permit the app to
run in background. It's not informative.

Due to the informative nature of "find out more", it's likely that the
user will not click it.

In order to improve the discoverability of the contents of the dialog,
this change embeds allow and forbid buttons in the notification.