RemoteDesktop: Bump to version 2 supporting session persistence
This is similar to the path in screencast. Both paths have to change as
a client in this session will call RemoteDesktop.SelectDevices and
ScreenCast.SetSources for the same session. Only the persist mode and
restore data for SelectDevices is honoured.

A prompt is shown if restore data was for a session without screens and
we try to restore screen as it changes the prompt the user selected.

A behavioural change is that the old code had a path for non-sandboxed
apps to get access regardless. This has been removed in leiu of the
persistence setting.