Default to allowing background applications without prompts
Flatpak tends to have a lot of applications with window IDs that do not
match the containment, particularly with X11 applications.

There is also the UX question of why we shouldn't an app run in the
background, KDE has support for systemtrays with both X11, and the FD.O
standard which most apps running in the background would populate.

This patch always returns "Allow once" when prompted by the daemon.

Whilst this might seem like a lazy solution, from what I can tell from a
lot of DBus montioring on Gnome, this is what they do too.

We can't just disable the portal implementation as disabling by the
config file doesn't work, but even if did we still want enableAutostart
to work.

I left the original code behind a configuration option in case we want
it for mobile.

BUG: 473881