remotedesktop: Don't ask for permission for non-sandboxed applications
Before we had restore support we had a check that would allow
non-sandboxed applications to access remote desktop without requiring
confirmation. This restores that check, but also enhances it, since it
often did not work properly. This was due to the app_id we receive not
being empty but the ID of the parent application. So now we not only
check if the app ID is empty, we also check if the application should be
considered sandboxed.