Set timeout on DBus calls to KWin
The new libei path in xwayland has a path where it can all deadlock each
other. Kwin makes blocking X calls to xwayland, xwayland makes a blocking call
to xdg-desktop-portal, and xdg-desktop-portal can't reply before getting a
response from kwin.

This is fixed in libei at:
https://gitlab.freedesktop.org/libinput/libei/-/merge_requests/295

For safety in the interim, we can set a timeout in the call from the
xdg-desktop-portal to kwin. If kwin takes more than 3 seconds, something is
clearly wrong and we're blocked somewhere. But it's also a short enough period
that a user won't have restarted or the systemd's watchdog kicked in.