Notification: Allow dbus-activating applications that are not running
This commit removes an isValid() check that will always return false when the target
application is not running, even though that application may exist, be activatable,
and implement the required interface while running.

By removing this check, we allow dbus to activate the application, which in turn
will appropriately deliver the dbus messages generated by interacting with the
notification.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>