[AppChooser] Correctly set m_autoRemember
Depending on whether the dialog is created from the portal API
or the private API for the KIO openwith dialog we want to automatically
remember the selection or not. In the portal case we want it to remember,
but that is missing, causing it to not be remembered.

To make the code more robust remove the default argument.

While at it also remove the unused parent arg