Implement the clipboard portal
Due to the async nature of the API we have to unfortunately use
event loop in order to use KSystemClipboard. This is because
QMimeData::retrieveData is a syncronous API and we have to
dispatch DBus method calls to communicate with the app while
in this function. If this turns out to be problematic in the
future, we can either move ClipboardPortal into a separate thread
and let it do it's thing there and block the main thread (for a
short time). Or implement the data control protocol directly.