ScreenChooserDialog: make items keyboard-accessible
The dialog for choosing a screen or window to share contains a
CardLayout  filled by a Repeater showing the screens or windows. This
works for pointer-based interaction, but does not allow keyboard
interaction; the cards function as buttons or checkboxes, but do not
accept keyboard focus so cannot be reached through the tab chain.

This change registers the cards in the tab chain, and highlights the
currently focused item.

CCBUG: 497971