ScreenChooserDialog: Set Accessible.role for Cards
The Cards for selecting screens or windows to share currently do not
explicitly set an Accessible.role, and screen readers will not inform
users about the type of control that they are interacting with.

This change sets the Accessible.role to Button for the Cards in single-
selection  dialogs (as activating them will immediately accept that
selection,  making it work like a button) and to CheckBox  for multiple-
selection dialogs, as activating them will only toggle the selected
status, requiring the use of the Share button to complete the selection.