RemoteDesktop: Send keysyms via fakeinput
We try to map back the keysym to a keycode. However that code is flawed.

We try to get the keyboard layout from an xkb state via xkb_state_serialize_layout.
However that functions is meant for xkb servers, which we aren't, so the whole
state logic doesn't work.

We also cannot reliably get the layout from the Wayland keyboard_modifiers event
since we are a background process.

Fix and simplify all of this by passing the keysym directly to KWin

BUG: 489021