Improve the Account access dialog
This dialog has a couple of usability problems that are immediately apparent:
The reason (if given) is haphazardly thrown at the end, which isn't very nice.
The user name and real name of the account is duplicated, and it's also not in
line with your avatar which makes it look odd. More importantly, the app name is
not displayed *anywhere* so it's impossible for the user to audit who is
creating this dialog - something we constantly get complaints about with our
other access dialogs.

This change hopes to address all of these issues, by re-arranging the dialog and
adding the requesting application name. The dialog is still functionally the
same as before.