kirigami-filepicker: Fix state reload
Since we don't recreate the file picker QML component with subsequent
FileOpen and FileSave calls, we need to ensure that the file picker's
state is properly reset. Currently, the file picker page makes some
assumptions that cause it to still have lingering state from previous
requests when a new request comes in.

This reinitializes the file picker page when a new request comes in.