InhibitPortal: Implement SessionMonitor
This is used for applications to inhibit logout.

The design is that plasma-shutdown will call into here to query if log
out is possible; this will in turn ask all applications if it is ok to
quit.

This makes it slightly different from other portals in that we have
non-sandboxed parts calling into here.

This patch also fixes a potential race with power inhibition if a caller
closes the request super early.