filechooser: correctly handle errors during fusing
it turns out dbus messages are like unions, when you look at the message
as a string it may in fact be the error string. put another way: by
looking at the return value (as a stringy value) you cannot tell if the
message was an error or not.

with that in mind we no longer coerce to stringy values but instead
check the error state and only use the value if there was no error.

this prevents us from accidentally getting past `if (foo.isempty)`
checks because foo is an error message.

example dbus message illustration

```
‣ Type=error  Endian=l  Flags=1  Version=1 Cookie=12  ReplyCookie=2
Timestamp="Mon 2025-04-07 09:55:53.476752 UTC"
  Sender=:1.318  Destination=:1.333
  ErrorName=org.kde.demo  ErrorMessage="Bar Error"
  UniqueName=:1.318
  MESSAGE "s" {
          STRING "Bar Error";
  };
```