fix: avoid empty applicationName in remote control dialog
This avoids having to display "An application has requested remote
control" in remote control dialog in case an application does not have a
desktop file. "An application" requesting permission makes it extremely
difficult to judge for the user what is going on.

I could only find out that `org.kde.krdpserver` was requesting
permission after applying this change. `org.kde.krdpserver` was
requesting this permission on every start-up but I didn't know which
application it actual was base on the previous dialog.