appchooser: Adjust text for "always open with"
This text makes a promise that we can't keep.

The xdg-desktop-portal effectively ignores the default for a lot of
situations, it has it's own logic to avoid the prompt if you select the
same app to open a file 3 times in a row.

This can be tracked at:
https://github.com/flatpak/xdg-desktop-portal/issues/472

In addition even if upstream behaviour is changed, the application can
force the "ask" flag which will show this dialog, or there's a path in
the permission store to always ask.

Setting the default still matters as it will be first in our list.

BUG: 503326


(cherry picked from commit c573eedb13ba04cd43960e81527bc386c96a8463)

75b89f91 appchooser: Adjust text for "always open with"
68e4dea3 Apply 1 suggestion(s) to 1 file(s)

Co-authored-by: David Edmundson <kde@davidedmundson.co.uk>