inhibit: Fix races between requests and calls to powerdevil
The request should exist as soon as possible so not Close
call is missed. Do not make the request context arg for
for the release call it can be deleted before the  return
is handled. Make sure the inhibition is released if the
request is closed while the inhibition call did not return.
Simplifies the release of the inhibition as well, clients
are unlikely to (be able to) handle Request::Close failing.
This allows Request::handleClose to be dropped.