userinfodialog: always set an avatar (from icon)
if the image is not pointing to an existing file then XDP will misbehave
and skip it from the payload, which runs counter to its own
documentation. simply put the file:// we passed in got dropped entirely.
indeed it's unclear what a client should do with file:// because the
documentation also says the uri will be that of a file.

long story short: instead of setting no avatar set the user icon as
avatar so we have something to pass to XDP

longer term we should think about always generating an avatar for the
user