globalshortcuts: Remember denied shortcuts
Currently if the user denies some shortcuts of an app on the next
start of the app it will likely try to bind these shortcuts again
resulting in the user seeing the dialog on every app start. To
prevent this shortcuts are now always registered (but not assigned
any keys if the user didn't want these shortcuts). This also enables
the user latter to assign key combinations to these actions in
systemsettings if they change their mind.