Avoid use of QDBusVirtualObject in Session / Change contstruction
QDBusVirtualObject is useful when interfaces are dynamic or paths are dynamic.
Neither of those are the case here and we use simpler object registration.

Move session registration to constructor, drop static creation

The static creation method works when Session code is all together, but
not when we want to put it with the rest of the code.

The connect to the Session close signal is not needed, the close method
which emits this already does the relevant cleanup.