port away from systemdialog
systemdialog is very cumbersome to make changes in and since it allows
look and feel packages to inject code it's nigh impossible to not break
things. also we generally consider this code injection a big no-no. it
is also not used outside XDPK.

instead let's use a Dialog implementation bespoke for XDPK. this gives
us more straight forward code and also allows us to more flexibly
iterate on the UI.

to that end this commit introduces a new PortalDialog element which is a
combination of SystemDialog.qmls in the component and the looknfeel.
for ease of porting no changes have been made besides those necessary to
support the merger of the two SystemDialog files.

in preparation of a general UI refresh the UI has changed some colors.