userinfodialog: reduce repeated calls into the dbus interface
previously we repeatedly read properties for no good reason. now we
don't because they all get cached in a UserDetails object

to enforce this pattern a bit more there's now a details object that
encapsulates the dbus interface. all data must be obtained from the
details object and as such is subject to caching