portaldialog: support adding items left of the button box
allows us to put contextually relevant options in the footer

OOTB buttonbox doesn't support adding things before the button view so
we need to turn the entire footer into a control. that better mimics the
header anyway though.

inside the footerControl we then have a layout with a leftcontainer and
the buttonbox as right container. for convergence reasons we use the new
flexboxlayout here so we can transparently reflow things while still
having layout control WRT filling (something Flow unfortunately doesn't
do -- alas, flexboxlayout is also not very reliable so equip the
containers with layout guides)