screenchooserdialog: add filter/search capability
this adds chips to filter specific screens and a search/filter field to
search for specific windows by name

the way this is implemented is through geometry intersection testing. a
window is relevant for a chip when its geometry intersects the chip's
output's geometry i.e. it is on that output, at least partially