PortalDialog: Fix various padding and sizing issues
This is especially noticeable in the "Remote control requested" dialog,
the padding between items in the header is wonky and icons tend to be
too large for its text.

I shrunk the size of the icon which better fits our heading, and also
increased the spacing between them. I also found that the spacing
between the header item and the window was wrong, it needed to be
largeSpacing according to our HIG. The result is a more balanced-feeling
window.